feat(product-service): move error-handler to layers
